@@ -24,28 +24,26 @@
     "url": "https://github.com/sitecore/jss/issues"
   },
   "devDependencies": {
-    "@types/jest": "^24.0.18",
-    "@vue/test-utils": "^1.0.0-beta.29",
+    "@types/jest": "^26.0.23",
+    "@vue/test-utils": "2.0.0-rc.6",
     "babel-core": "^6.26.3",
     "del-cli": "^3.0.1",
     "eslint": "^7.15.0",
-    "eslint-plugin-vue": "^7.3.0",
-    "jest": "^26.4.2",
+    "eslint-plugin-vue": "^7.9.0",
+    "jest": "^26.6.3",
     "jest-serializer-html": "^7.0.0",
-    "ts-jest": "24.1.0",
-    "typescript": "~3.6.3",
-    "vue": "^2.6.10",
-    "vue-class-component": "^7.1.0",
-    "vue-jest": "^3.0.5",
-    "vue-property-decorator": "^8.2.2",
-    "vue-server-renderer": "^2.6.12",
-    "vue-template-compiler": "^2.6.10"
+    "ts-jest": "26.5.6",
+    "typescript": "~4.2.4",
+    "vue": "^v3.1.0-beta.6",
+    "vue-class-component": "^7.2.6",
+    "vue-jest": "^5.0.0-alpha.9"
   },
   "peerDependencies": {
-    "vue": "^2.6.10"
+    "vue": "^v3.1.0-beta.6"
   },
   "dependencies": {
-    "@sitecore-jss/sitecore-jss": "^19.0.0-canary.4"
+    "@sitecore-jss/sitecore-jss": "^19.0.0-canary.4",
+    "@vue/compiler-sfc": "^3.0.11"
   },
   "types": "./types/index.d.ts",
   "gitHead": "2f4820efddf4454eeee58ed1b2cc251969efdf5b"
