@@ -48,25 +48,25 @@
     "lint": "vue-cli-service lint ./src/**/*.vue ./src/**/*.js ./sitecore/definitions/**/*.js ./scripts/**/*.js ./server/**/*.js ./data/**/*.yml"
   },
   "dependencies": {
+    "@apollo/client": "^3.3.19",
     "@panter/vue-i18next": "~0.15.1",
     "@sitecore-jss/sitecore-jss-vue": "^19.0.0-canary.4",
-    "apollo-cache-inmemory": "~1.6.3",
-    "apollo-client": "~2.6.4",
-    "apollo-link": "~1.2.13",
+    "@vue/apollo-composable": "^4.0.0-alpha.12",
+    "@vue/apollo-option": "^4.0.0-alpha.11",
+    "@vue/apollo-ssr": "^4.0.0-alpha.11",
+    "@vue/server-renderer": "^3.0.11",
     "apollo-link-batch-http": "~1.2.13",
     "apollo-link-persisted-queries": "~0.2.2",
     "axios": "^0.21.1",
     "bootstrap": "^4.3.1",
     "cross-fetch": "~2.2.3",
-    "graphql": "~14.5.8",
-    "i18next": "~17.0.16",
-    "i18next-fetch-backend": "~2.2.0",
+    "graphql": "^15.5.0",
     "register-service-worker": "~1.6.2",
-    "vue": "~2.6.10",
-    "vue-apollo": "3.0.0-beta.27",
-    "vue-meta": "~2.2.2",
-    "vue-router": "~3.1.3",
-    "vue-server-renderer": "^2.6.12"
+    "serialize-javascript": "^5.0.1",
+    "vue": "^v3.1.0-beta.6",
+    "vue-i18n": "^9.1.6",
+    "vue-meta": "3.0.0-alpha.6",
+    "vue-router": "^4.0.8"
   },
   "devDependencies": {
     "@babel/register": "7.6.2",
@@ -74,21 +74,21 @@
     "@sitecore-jss/sitecore-jss-dev-tools": "^19.0.0-canary.4",
     "@sitecore-jss/sitecore-jss-manifest": "^19.0.0-canary.4",
     "@sitecore-jss/sitecore-jss-tracking": "^19.0.0-canary.4",
-    "@vue/cli-plugin-babel": "~4.5.6",
-    "@vue/cli-plugin-eslint": "~4.5.6",
-    "@vue/cli-service": "~4.5.6",
+    "@vue/cli-plugin-babel": "~4.5.13",
+    "@vue/cli-plugin-eslint": "~4.5.13",
+    "@vue/cli-service": "~4.5.13",
+    "@vue/compiler-sfc": "^3.0.11",
     "@vue/eslint-config-prettier": "~5.0.0",
     "babel-eslint": "~10.0.3",
     "chokidar": "~3.1.1",
     "cross-env": "~6.0.0",
     "eslint": "^7.4.0",
     "eslint-plugin-prettier": "^3.1.3",
-    "eslint-plugin-vue": "~5.2.3",
+    "eslint-plugin-vue": "~7.9.0",
     "eslint-plugin-yaml": "^0.2.0",
-    "graphql-tag": "~2.10.1",
+    "graphql-tag": "^2.12.4",
     "html-loader": "~0.5.5",
     "npm-run-all": "~4.1.5",
-    "null-loader": "~3.0.0",
-    "vue-template-compiler": "~2.6.10"
+    "null-loader": "~3.0.0"
   }
 }
