@@ -1,5 +1,5 @@
 import { mediaApi } from '@sitecore-jss/sitecore-jss';
-import { CreateElement, FunctionalComponentOptions, RenderContext } from 'vue';
+import { h, defineComponent, PropType } from 'vue';
 import { generateHtmlTag } from '../utils';
 
 export interface ImageFieldValue {
@@ -13,40 +13,12 @@ export interface ImageField {
   editable?: string;
 }
 
-export interface ImageProps {
-  /** The image field data. */
-  media: ImageField | ImageFieldValue;
-
-  /**
-   * Can be used to explicitly disable inline editing.
-   * If true and `media.editable` has a value, then `media.editable` will be processed
-   * and rendered as component output. If false, `media.editable` value will be ignored and not rendered.
-   */
-  editable?: boolean;
-
-  /**
-   * Custom regexp that finds media URL prefix that will be replaced by `/-/jssmedia` or `/~/jssmedia`.
-   * @example
-   * /\/([-~]{1})assets\//i
-   * /-assets/website -> /-/jssmedia/website
-   * /~assets/website -> /~/jssmedia/website
-   */
-  mediaUrlPrefix?: RegExp;
-
-  /**
-   * Parameters that will be attached to Sitecore media URLs
-   */
-  imageParams?: {
-    [paramName: string]: string;
-  };
-}
-
-const getEditableWrapper = (editableMarkup: string, createElement: CreateElement) =>
+const getEditableWrapper = (editableMarkup: string) =>
   // create an inline wrapper and use dangerouslySetInnerHTML.
   // if we try to parse the EE value, the parser may strip invalid or disallowed attributes from html elements - and EE uses several
-  createElement('span', {
+  h('span', {
     class: 'sc-image-wrapper',
-    domProps: { innerHTML: editableMarkup },
+    innerHTML: editableMarkup,
   });
 
 const getImageAttrs = (
@@ -80,24 +52,46 @@ const getImageAttrs = (
   return newAttrs;
 };
 
-// custom comment to enable eslint vue plugin for Vue components as objects: https://github.com/vuejs/eslint-plugin-vue#attention
-// @vue/component
-export const Image: FunctionalComponentOptions<ImageProps> = {
-  functional: true,
+export const Image = defineComponent({
+  inheritAttrs: false,
   props: {
-    media: { type: Object, required: true },
+    /** The image field data. */
+    media: {
+      type: Object as PropType<ImageField | ImageFieldValue>,
+      default() {
+        return {} as ImageField | ImageFieldValue;
+      },
+      required: true,
+    },
+    /**
+     * Can be used to explicitly disable inline editing.
+     * If true and `media.editable` has a value, then `media.editable` will be processed
+     * and rendered as component output. If false, `media.editable` value will be ignored and not rendered.
+     */
     editable: { type: Boolean, default: true },
-    imageParams: { type: Object, default: () => ({}) },
+    /**
+     * Parameters that will be attached to Sitecore media URLs
+     */
+    imageParams: {
+      type: Object as PropType<{
+        [paramName: string]: string;
+      }>,
+      default: () => ({}),
+    },
+    /**
+     * Custom regexp that finds media URL prefix that will be replaced by `/-/jssmedia` or `/~/jssmedia`.
+     * @example
+     * /\/([-~]{1})assets\//i
+     * /-assets/website -> /-/jssmedia/website
+     * /~assets/website -> /~/jssmedia/website
+     */
     mediaUrlPrefix: { type: RegExp, default: undefined },
   },
-  // Need to assign `any` return type because Vue type definitions are inaccurate.
-  // The Vue type definitions set `render` to a return type of VNode and that's it.
-  // However, it is possible to return null | string | VNode[] | VNodeChildrenArrayContents.
-  render(createElement: CreateElement, context: RenderContext): any {
-    const { media, editable, imageParams, mediaUrlPrefix } = context.props;
-    const contextAttrs = context.data.attrs;
+  render() {
+    const { media, editable, imageParams, mediaUrlPrefix } = this.$props;
+    const contextAttrs = this.$attrs;
 
-    if (!media || (!media.editable && !media.value && !media.src)) {
+    if (!media || (!media.editable && !media.value && !(media as ImageFieldValue).src)) {
       return null;
     }
 
@@ -105,7 +99,7 @@ export const Image: FunctionalComponentOptions<ImageProps> = {
     if (editable && media.editable) {
       const foundImg = mediaApi.findEditorImageTag(media.editable);
       if (!foundImg) {
-        return getEditableWrapper(media.editable, createElement);
+        return getEditableWrapper(media.editable);
       }
 
       const imgAttrs = getImageAttrs(
@@ -114,29 +108,29 @@ export const Image: FunctionalComponentOptions<ImageProps> = {
         mediaUrlPrefix
       );
       if (!imgAttrs) {
-        return getEditableWrapper(media.editable, createElement);
+        return getEditableWrapper(media.editable);
       }
 
       const imgHtml = generateHtmlTag('img', imgAttrs);
       const editableMarkup = media.editable.replace(foundImg.imgTag, imgHtml);
-      return getEditableWrapper(editableMarkup, createElement);
+      return getEditableWrapper(editableMarkup);
     }
 
     // some wise-guy/gal is passing in a 'raw' image object value
-    const img = media.src ? media : media.value;
+    const img = (media as ImageFieldValue).src ? media : media.value;
     if (!img) {
       return null;
     }
 
     const attrs = getImageAttrs({ ...img, ...contextAttrs }, imageParams, mediaUrlPrefix);
     if (attrs) {
-      // in functional components, context.data should be passed along to the
+      // this.$data should be passed along to the
       // `createElement` function in order to retain attributes and events
-      // https://vuejs.org/v2/guide/render-function.html#Passing-Attributes-and-Events-to-Child-Elements-Components
-      const data = { ...context.data, attrs };
-      return createElement('img', data);
+      // https://v3.vuejs.org/guide/render-function.html#render-functions
+      const data = { ...this.$data, ...attrs };
+      return h('img', data);
     }
 
     return null; // we can't handle the truth
   },
-};
+});
