@@ -1,4 +1,4 @@
-import { CreateElement, FunctionalComponentOptions, RenderContext } from 'vue';
+import { defineComponent, h, PropType } from 'vue';
 
 export interface LinkFieldValue {
   [attributeName: string]: any;
@@ -14,48 +14,38 @@ export interface LinkField {
   editableLastPart?: string;
 }
 
-export interface LinkProps {
-  /** The link field data. */
-  field: LinkField | LinkFieldValue;
-  /**
-   * Can be used to explicitly disable inline editing.
-   * If true and `field.editable` has a value,
-   * then `field.editable` will be processed and rendered as
-   * component output. If false, `field.editable` value will be ignored and not rendered.
-   */
-  editable?: boolean;
-
-  /**
-   * Displays a link text ('description' in Sitecore) even when children exist
-   * NOTE: when in Sitecore Experience Editor, this setting is ignored due to technical limitations, and the description is always rendered.
-   */
-  showLinkTextWithChildrenPresent?: boolean;
-}
-
-export const Link: FunctionalComponentOptions<LinkProps> = {
-  functional: true,
+export const Link = defineComponent({
   props: {
+    /** The link field data. */
     field: {
-      type: Object,
+      type: Object as PropType<LinkField | LinkFieldValue>,
+      default() {
+        return {} as LinkField | LinkFieldValue;
+      },
       required: true,
     },
+    /**
+     * Can be used to explicitly disable inline editing.
+     * If true and `field.editable` has a value,
+     * then `field.editable` will be processed and rendered as
+     * component output. If false, `field.editable` value will be ignored and not rendered.
+     */
     editable: {
       type: Boolean,
       default: true,
     },
+    /**
+     * Displays a link text ('description' in Sitecore) even when children exist
+     * NOTE: when in Sitecore Experience Editor, this setting is ignored due to technical limitations, and the description is always rendered.
+     */
     showLinkTextWithChildrenPresent: {
       type: Boolean,
       default: false,
     },
   },
-  // Need to assign `any` return type because Vue type definitions are inaccurate.
-  // The Vue type definitions set `render` to a return type of VNode and that's it.
-  // However, it is possible to return null | string | VNode[] | VNodeChildrenArrayContents.
-  render(createElement: CreateElement, context: RenderContext): any {
-    const {
-      children,
-      props: { field, editable, showLinkTextWithChildrenPresent },
-    } = context;
+  render() {
+    const { field, editable, showLinkTextWithChildrenPresent } = this.$props;
+    const children = this.$slots.default;
 
     const dynamicField: any = field;
 
@@ -75,28 +65,21 @@ export const Link: FunctionalComponentOptions<LinkProps> = {
 
       markup += dynamicField.editableLastPart;
 
-      // in functional components, context.data should be passed along to the
+      // this.$data should be passed along to the
       // `createElement` function in order to retain attributes and events
-      // https://vuejs.org/v2/guide/render-function.html#Passing-Attributes-and-Events-to-Child-Elements-Components
+      // https://v3.vuejs.org/guide/render-function.html#render-functions
       const elementData = {
-        ...context.data,
+        ...this.$data,
         class: 'sc-link-wrapper',
-        domProps: { innerHTML: markup },
+        innerHTML: markup,
       };
 
-      const xEditorElement = createElement('span', elementData);
+      const xEditorElement = h('span', elementData);
 
       if (children) {
-        const childElements = createElement(
-          'span',
-          { class: 'sc-link-editable-children-wrapper' },
-          children
-        );
+        const childElements = h('span', { class: 'sc-link-editable-children-wrapper' }, children());
 
-        return createElement('span', { class: 'sc-link-editable-wrapper' }, [
-          xEditorElement,
-          childElements,
-        ]);
+        return h('span', { class: 'sc-link-editable-wrapper' }, [xEditorElement, childElements]);
       }
 
       return xEditorElement;
@@ -113,16 +96,19 @@ export const Link: FunctionalComponentOptions<LinkProps> = {
         ? link.text || link.href
         : null;
 
-    const finalChildren = children ? [linkText, ...children] : linkText;
+    const finalChildren = children ? [linkText, ...children()] : linkText;
 
-    // in functional components, context.data should be passed along to the
+    // this.$data should be passed along to the
     // `createElement` function in order to retain attributes and events
-    // https://vuejs.org/v2/guide/render-function.html#Passing-Attributes-and-Events-to-Child-Elements-Components
+    // https://v3.vuejs.org/guide/render-function.html#render-functions
     const data = {
-      ...context.data,
+      ...this.$data,
       class: link.class,
-      attrs: { ...context.data.attrs, href: link.href, title: link.title, target: link.target },
+      ...this.$attrs,
+      href: link.href,
+      title: link.title,
+      target: link.target,
     };
-    return createElement('a', data, finalChildren);
+    return h('a', data, finalChildren);
   },
-};
+});
