@@ -1,26 +1,26 @@
-import { CreateElement, FunctionalComponentOptions, RenderContext } from 'vue';
+import { h, defineComponent, PropType } from 'vue';
 
-export interface MissingComponentProps {
-  rendering?: {
-    componentName?: string;
-  };
-}
-
-export const MissingComponent: FunctionalComponentOptions<MissingComponentProps> = {
-  functional: true,
+export const MissingComponent = defineComponent({
   props: {
     rendering: {
-      type: Object,
+      type: Object as PropType<{
+        componentName?: string;
+      }>,
+      default() {
+        return undefined as {
+          componentName?: string;
+        };
+      },
     },
   },
-  render(createElement: CreateElement, context: RenderContext) {
-    const { rendering } = context.props;
+  render() {
+    const { rendering } = this.$props;
     const componentName =
       rendering && rendering.componentName ? rendering.componentName : 'Unnamed Component';
 
-    console.log(`Component props for unimplemented '${componentName}' component`, context.props);
+    console.log(`Component props for unimplemented '${componentName}' component`, this.$props);
 
-    return createElement(
+    return h(
       'div',
       {
         style: {
@@ -32,12 +32,12 @@ export const MissingComponent: FunctionalComponentOptions<MissingComponentProps>
         },
       },
       [
-        createElement('h2', componentName),
-        createElement(
+        h('h2', componentName),
+        h(
           'p',
           'JSS component is missing Vue implementation. See the developer console for more information.'
         ),
       ]
     );
   },
-};
+});
