@@ -6,7 +6,7 @@ import {
   Item,
   RouteData,
 } from '@sitecore-jss/sitecore-jss';
-import { Component, CreateElement, VNode } from 'vue';
+import { Component, h, VNode, DefineComponent } from 'vue';
 import { MissingComponent } from './MissingComponent';
 import { ComponentFactory } from './sharedTypes';
 
@@ -40,13 +40,13 @@ export interface PlaceholderProps {
    * A component that is rendered in place of any components that are in this placeholder,
    * but do not have a definition in the componentFactory (i.e. don't have a React implementation)
    */
-  missingComponentComponent?: Component;
+  missingComponentComponent?: DefineComponent;
 
   /**
    * A component that is rendered in place of the placeholder when an error occurs rendering
    * the placeholder
    */
-  errorComponent?: Component;
+  errorComponent?: DefineComponent;
 }
 
 export type JssDynamicComponent = Component & { isxEditorComponent?: boolean };
@@ -81,14 +81,12 @@ export const getPlaceholderDataFromRenderingData = (
  * @see convertVNodesToDynamicComponents or @see getDynamicComponentsFromRenderingData for options.
  * @param {Array<ComponentRendering | HtmlElementRendering>} placeholderData
  * @param {PlaceholderProps} placeholderProps
- * @param {CreateElement} createVueElement
  * @param {ComponentFactory} [componentFactory]
  * @returns {VNode[]} vnodes
  */
 export function getVNodesForRenderingData(
   placeholderData: Array<ComponentRendering | HtmlElementRendering>,
   placeholderProps: PlaceholderProps,
-  createVueElement: CreateElement,
   componentFactory?: ComponentFactory
 ) {
   const {
@@ -105,10 +103,10 @@ export function getVNodesForRenderingData(
 
       // if the element is not a 'component rendering', render it 'raw'
       if (!rendering.componentName && rendering.name) {
-        return createRawElement(rendering, createVueElement);
+        return createRawElement(rendering);
       }
 
-      let component = getComponentForRendering(rendering, componentFactory);
+      let component: any = getComponentForRendering(rendering, componentFactory);
       if (!component) {
         console.error(
           `Placeholder ${placeholderName} contains unknown component ${rendering.componentName}. Ensure that a Vue component exists for it, and that it is mapped in your component factory.`
@@ -125,7 +123,7 @@ export function getVNodesForRenderingData(
         finalProps.params = { ...placeholderParams, ...rendering.params };
       }
 
-      return createVueElement(component, { props: finalProps, key });
+      return h(component, { ...finalProps, key });
     })
     .filter((element) => element) as VNode[]; // remove nulls;
 }
@@ -135,23 +133,21 @@ export function getVNodesForRenderingData(
  * to return "renderable" components, i.e. components that can be rendered in a Vue template, a.k.a. dynamic components.
  * @param {Array<ComponentRendering | HtmlElementRendering>} placeholderData
  * @param {PlaceholderProps} placeholderProps
- * @param {CreateElement} createVueElement
  * @param {ComponentFactory} componentFactory
  * @returns {JssDynamicComponent[]} dynamic components
  */
 export function getDynamicComponentsFromRenderingData(
   placeholderData: Array<ComponentRendering | HtmlElementRendering>,
   placeholderProps: PlaceholderProps,
-  createVueElement: CreateElement,
   componentFactory?: ComponentFactory
 ) {
   return convertVNodesToDynamicComponents(
-    getVNodesForRenderingData(placeholderData, placeholderProps, createVueElement, componentFactory)
+    getVNodesForRenderingData(placeholderData, placeholderProps, componentFactory)
   );
 }
 
 /**
- * Converts VNodes to simple functional components that render the vnode.
+ * Converts VNodes to components that render the vnode.
  * Also evaluates VNodes to determine if they are Experience Editor components/elements and if so,
  * adds an identifying property to the component.
  * @param {VNode[]} vnodes
@@ -160,18 +156,15 @@ export function getDynamicComponentsFromRenderingData(
 export function convertVNodesToDynamicComponents(vnodes: VNode[]) {
   return vnodes.map((vnode) => {
     const component = {
-      functional: true,
-      props: vnode.data && vnode.data.props,
+      $props: vnode.props,
+      inheritAttrs: false,
+
       render() {
         return vnode;
       },
     } as JssDynamicComponent;
-    if (
-      vnode.tag === 'code' &&
-      vnode.data &&
-      vnode.data.attrs &&
-      vnode.data.attrs.type === 'text/sitecore'
-    ) {
+
+    if (vnode.type === 'code' && vnode.props.type === 'text/sitecore') {
       component.isxEditorComponent = true;
     }
     return component;
@@ -180,9 +173,8 @@ export function convertVNodesToDynamicComponents(vnodes: VNode[]) {
 
 /**
  * @param {any} elem
- * @param {CreateElement} createVueElement
  */
-function createRawElement(elem: any, createVueElement: CreateElement) {
+function createRawElement(elem: any) {
   if (!elem.name) {
     console.error(
       '"elem.name" is undefined in "createRawElement". ' +
@@ -191,13 +183,7 @@ function createRawElement(elem: any, createVueElement: CreateElement) {
     return null;
   }
 
-  const attrs = elem.attributes;
-
-  const domProps = {
-    innerHTML: elem.contents,
-  };
-
-  const component = createVueElement(elem.name, { attrs, domProps });
+  const component = h(elem.name, { ...elem.attributes, innerHTML: elem.contents });
 
   return component;
 }
