@@ -12,7 +12,7 @@ describe('<Text />', () => {
     const errorSpy = jest.spyOn(console, 'error');
     errorSpy.mockImplementation(() => {});
     const rendered = mount(Text);
-    expect(rendered.isEmpty()).toBe(true);
+    expect(rendered.element.innerHTML).toBe(undefined);
     errorSpy.mockRestore();
   });
 
@@ -22,8 +22,8 @@ describe('<Text />', () => {
     // that is marked as an Object.
     const errorSpy = jest.spyOn(console, 'error');
     errorSpy.mockImplementation(() => {});
-    const rendered = mount(Text, { context: { props } });
-    expect(rendered.isEmpty()).toBe(true);
+    const rendered = mount(Text, { props });
+    expect(rendered.element.innerHTML).toBe(undefined);
     errorSpy.mockRestore();
   });
 
@@ -37,8 +37,8 @@ describe('<Text />', () => {
     // that is marked as an Object.
     const errorSpy = jest.spyOn(console, 'error');
     errorSpy.mockImplementation(() => {});
-    const rendered = mount(Text, { context: { props } });
-    expect(rendered.isEmpty()).toBe(true);
+    const rendered = mount(Text, { props });
+    expect(rendered.element.innerHTML).toBe(undefined);
     errorSpy.mockRestore();
   });
 
@@ -49,7 +49,7 @@ describe('<Text />', () => {
         editable: 'editable',
       },
     };
-    const rendered = mount(Text, { context: { props } }).find('span');
+    const rendered = mount(Text, { props }).find('span');
     expect(rendered.exists()).toBe(true);
     expect(rendered.element.innerHTML).toBe(props.field.editable);
   });
@@ -62,7 +62,7 @@ describe('<Text />', () => {
       },
       editable: false,
     };
-    const rendered = mount(Text, { context: { props } }).find('span');
+    const rendered = mount(Text, { props }).find('span');
     expect(rendered.exists()).toBe(true);
     expect(rendered.element.innerHTML).toBe(props.field.value);
   });
@@ -74,7 +74,7 @@ describe('<Text />', () => {
       },
       editable: false,
     };
-    const rendered = mount(Text, { context: { props } }).find('span');
+    const rendered = mount(Text, { props }).find('span');
     expect(rendered.exists()).toBe(true);
     expect(rendered.html().indexOf('&lt; &gt;')).toBeGreaterThan(-1);
   });
@@ -85,7 +85,7 @@ describe('<Text />', () => {
         value: 'value',
       },
     };
-    const rendered = mount(Text, { context: { props } }).find('span');
+    const rendered = mount(Text, { props }).find('span');
     expect(rendered.exists()).toBe(true);
     expect(rendered.element.innerHTML).toBe(props.field.value);
   });
@@ -96,7 +96,7 @@ describe('<Text />', () => {
         value: 1.23,
       },
     };
-    const rendered = mount(Text, { context: { props } }).find('span');
+    const rendered = mount(Text, { props }).find('span');
     expect(rendered.exists()).toBe(true);
     expect(rendered.element.innerHTML).toBe(props.field.value.toString());
   });
@@ -107,7 +107,7 @@ describe('<Text />', () => {
         value: 0,
       },
     };
-    const rendered = mount(Text, { context: { props } }).find('span');
+    const rendered = mount(Text, { props }).find('span');
     expect(rendered.exists()).toBe(true);
     expect(rendered.element.innerHTML).toBe(props.field.value.toString());
   });
@@ -119,7 +119,7 @@ describe('<Text />', () => {
       },
       encode: false,
     };
-    const rendered = mount(Text, { context: { props } }).find('span');
+    const rendered = mount(Text, { props }).find('span');
     expect(rendered.exists()).toBe(true);
     expect(rendered.element.innerHTML).toBe(props.field.value);
   });
@@ -130,7 +130,7 @@ describe('<Text />', () => {
         editable: eeTextData,
       },
     };
-    const rendered = mount(Text, { context: { props } }).find('span');
+    const rendered = mount(Text, { props }).find('span');
     expect(rendered.exists()).toBe(true);
     expect(rendered.html().indexOf('<input')).toBeGreaterThan(-1);
     expect(rendered.html().indexOf('<span class="scChromeData">')).toBeGreaterThan(-1);
@@ -143,7 +143,7 @@ describe('<Text />', () => {
       },
       tag: 'h1',
     };
-    const rendered = mount(Text, { context: { props } }).find('h1');
+    const rendered = mount(Text, { props }).find('h1');
     expect(rendered.exists()).toBe(true);
     expect(rendered.element.innerHTML).toBe(props.field.value);
   });
@@ -159,7 +159,7 @@ describe('<Text />', () => {
       class: 'cssClass',
       id: 'lorem',
     };
-    const rendered = mount(Text, { context: { props, attrs } }).find('h1');
+    const rendered = mount(Text, { props, attrs }).find('h1');
     expect(rendered.exists()).toBe(true);
     expect(rendered.element.tagName).toBe(props.tag.toUpperCase());
     expect(rendered.attributes()).toMatchObject(attrs);
