@@ -1,5 +1,8 @@
 <template>
-  <not-found v-if="notFound" :context="appState.sitecoreContext" />
+  <not-found
+    v-if="notFound && !loading && !languageIsChanging"
+    :context="appState.sitecoreContext"
+  />
   <route-loading v-else-if="loading" />
   <layout v-else :route="appState.routeData" />
 </template>
@@ -26,7 +29,7 @@ export default {
 
     // To take advantage of Vue's reactive data for tracking app state changes, we need
     // to reference the same `state` object that the $jss store references in order for mutations to be observed.
-    // $jss is attached to the Vue instance via `SitecoreJssPlugin`.
+    // $jss is attached to the App instance via `SitecoreJssPlugin`.
     const appState = this.$jss.store.state;
 
     // if the app state has routeData, we don't need to load it and don't need a loading screen
@@ -68,6 +71,9 @@ export default {
     this.updateLanguage();
   },
   inject: {
+    languageIsChanging: {
+      type: Boolean,
+    },
     changeAppLanguage: {
       type: Function,
     },
@@ -77,7 +83,9 @@ export default {
      * Loads route data from Sitecore Layout Service into appState.routeData
      */
     updateRouteData() {
-      let sitecoreRoutePath = this.route.params.sitecoreRoute || '/';
+      let sitecoreRoutePath = this.route.params.sitecoreRoute
+        ? this.route.params.sitecoreRoute.join('/')
+        : '/';
       if (!sitecoreRoutePath.startsWith('/')) {
         sitecoreRoutePath = `/${sitecoreRoutePath}`;
       }
