@@ -22,16 +22,10 @@ const store = {
 };
 
 // Vue plugins must export a function named 'install'
-function install(Vue) {
-  // "standard" convention for Vue plugins to ensure they are only installed once.
-  if (install.installed) {
-    return;
-  }
-  install.installed = true;
-
-  Vue.prototype.$jss = {
+function install(App) {
+  App.config.globalProperties.$jss = {
     // there may be other JSS plugins installed, merge existing properties
-    ...Vue.prototype.$jss,
+    ...(App.config.globalProperties.$jss || {}),
     store,
     sitecoreContext() {
       // this is intended only as a convenience function for easier access to the current context.
